+++ lib/features/exercises/data/datasources/exercises_datasource.dart
@@ -1,0 +1,29 @@
+import 'package:pop_network/pop_network.dart';
+
+class ExercisesDatasource {
+  final IApiManager _apiManager;
+
+  ExercisesDatasource(this._apiManager);
+
+  // Note que usamos o tipo de retorno da biblioteca
+  Future<Response<dynamic>> getExercises({
+    String? name,
+    String? type,
+    String? muscle,
+    String? difficulty,
+    List<String>? equipments,
+  }) => _apiManager.get(
+    '/v1/exercises',
+    queryParameters: {
+      if (name != null) 'name': name,
+      if (type != null) 'type': type,
+      if (muscle != null) 'muscle': muscle,
+      if (difficulty != null) 'difficulty': difficulty,
+      if (equipments?.isNotEmpty ?? false) 'equipment': equipments!.first,
+    },
+    mockReplyParams: MockReplyParams(
+      mockPath: 'exercises_list_mock',
+      status: HttpStatusEnum.ok,
+    ),
+  );
+}