+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 * feat: implementação da camada Domain (Entities, Repository Interface e UseCases)
 * test: adição de testes unitários do GetExercisesUseCase (sucesso e erro)
+
+## 0.2.0
+
+* feat: implementação da camada Data (Models, DataSources e Repositories Implementation)
+* test: adição de testes unitários para o Repository e DataSources